implement side panel flyouts
