+++ express/code/blocks/font-generator/font-generator.css
@@ -2,21 +2,59 @@
   display: none;
 }
 
+/*
+ * Up to 1440px the layout is a single column: the side panel (preview input +
+ * suggestions) sits full-width on top of the results below it. At 1440px+ it
+ * becomes the 12-column grid with the side panel as a 4-col rail.
+ */
 .font-generator-grid {
   display: grid;
-  grid-template-columns: repeat(12, minmax(0, 1fr));
+  grid-template-columns: 1fr;
   gap: var(--spacing-300, 16px);
   background: var(--gray-75, #F3F3F3);
 }
 
-.font-generator-col--side {
-  grid-column: span 4;
+.font-generator-col {
+  min-width: 0;
 }
 
-.font-generator-col--main {
-  grid-column: span 4;
+/* Filter trigger — only shown while filters live in the sliding panel (<1440px). */
+.font-generator-filter-trigger {
+  align-self: flex-start;
+  margin: 0 40px;
+  padding: 8px 20px;
+  border: 1px solid var(--s2a-color-s2ac-palette-gray-100, #e9e9e9);
+  border-radius: 24px;
+  background-color: #ffffff;
+  color: var(--theme-alias-content-typography-title, #131313);
+  font-family: var(--title-s-font-family, "Adobe Clean Spectrum VF", Helvetica);
+  font-size: 14px;
+  font-weight: 700;
+  cursor: pointer;
 }
 
-.font-generator-col--aux {
-  grid-column: span 4;
+.font-generator-filter-trigger:hover {
+  background-color: var(--s2a-color-s2ac-palette-gray-100, #e9e9e9);
+}
+
+@media (min-width: 1440px) {
+  .font-generator-grid {
+    grid-template-columns: repeat(12, minmax(0, 1fr));
+  }
+
+  .font-generator-col--side {
+    grid-column: span 4;
+  }
+
+  .font-generator-col--main {
+    grid-column: span 4;
+  }
+
+  .font-generator-col--aux {
+    grid-column: span 4;
+  }
+
+  .font-generator-filter-trigger {
+    display: none;
+  }
 }
