+++ express/code/blocks/font-generator/font-generator.js
@@ -1,4 +1,11 @@
+import { getLibs } from '../../scripts/utils.js';
 import { createSidePanel } from './side-panel/output/side-panel.js';
+import { createFilterPanel } from './side-panel/output/filter-panel.js';
+import { getState, setState, subscribe } from './state.js';
+
+const CATEGORY_STYLES = { all: 'bold-script', glitch: 'strikethrough', 'symbol text': 'weights' };
+
+let filterPanelCount = 0;
 
 // Reads the side-panel copy from the authored block. The first paragraph holds
 // comma-separated preview suggestions; the link is the promo CTA; the remaining
@@ -24,6 +31,46 @@ function extractContent(block) {
   };
 }
 
+// Async-enhances the trigger label from placeholders so it stays translatable;
+// keeps the authored fallback if the 'filter' placeholder is not defined.
+async function localizeFilterTrigger(button) {
+  try {
+    const { getConfig } = await import(`${getLibs()}/utils/utils.js`);
+    const { replaceKey } = await import(`${getLibs()}/features/placeholders.js`);
+    const label = await replaceKey('filter', getConfig());
+    if (label) {
+      button.textContent = label;
+      button.setAttribute('aria-label', label);
+    }
+  } catch (e) {
+    // Placeholder lookup unavailable — keep the fallback label.
+  }
+}
+
+// The Filter trigger lives outside the side-panel component. It toggles the
+// shared filter panel through state (state.filtersOpen); the panel subscribes
+// to that flag for its open/closed slide. Hidden via CSS at >=1440px where
+// filters are inline.
+function createFilterTrigger(panelId) {
+  const button = document.createElement('button');
+  button.type = 'button';
+  button.className = 'font-generator-filter-trigger';
+  button.textContent = 'Filter';
+  button.setAttribute('aria-haspopup', 'true');
+  button.setAttribute('aria-expanded', 'false');
+  button.setAttribute('aria-controls', panelId);
+
+  button.addEventListener('click', () => {
+    setState({ filtersOpen: !getState().filtersOpen });
+  });
+  subscribe(({ filtersOpen }) => {
+    button.setAttribute('aria-expanded', String(Boolean(filtersOpen)));
+  });
+
+  localizeFilterTrigger(button);
+  return button;
+}
+
 export default function decorate(block) {
   block.classList.add('loading');
 
@@ -34,10 +81,20 @@ export default function decorate(block) {
 
   const sideCol = document.createElement('div');
   sideCol.className = 'font-generator-col font-generator-col--side';
-  sideCol.append(createSidePanel({
-    ...content,
-    categoryStyles: { all: 'bold-script', 'symbol text': 'weights' },
-  }));
+
+  const panelId = `font-generator-filters-${(filterPanelCount += 1)}`;
+  const filterPanel = createFilterPanel({
+    promoTitle: content.promoTitle,
+    promoCta: content.promoCta,
+    categoryStyles: CATEGORY_STYLES,
+  });
+  filterPanel.id = panelId;
+
+  sideCol.append(
+    createSidePanel({ suggestions: content.suggestions }),
+    createFilterTrigger(panelId),
+    filterPanel,
+  );
 
   const mainCol = document.createElement('div');
   mainCol.className = 'font-generator-col font-generator-col--main';
